requirements.txt
asgiref==3.3.4
Django==3.2
django-taggit==1.5.0
gunicorn==20.1.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
whitenoise==5.2.0
django-autoslug==1.9.8
django-countries==7.2.1
django-countries-plus==1.3.1